change event triggers to CARDINALITY.ONE, as there seems to be an odd thing happening when deployed to azure
