+++ fns-csv-pipeline/src/test/kotlin/gov/cdc/dex/csv/functions/Unit_FnDecompressor.kt
@@ -19,7 +19,6 @@ import org.mockito.invocation.InvocationOnMock;
 
 import gov.cdc.dex.csv.services.BlobService
 import gov.cdc.dex.csv.services.EventService
-import gov.cdc.dex.csv.test.utils.*
 import gov.cdc.dex.csv.dtos.ConnectionNames
 import gov.cdc.dex.csv.dtos.EventHubNames
 import gov.cdc.dex.csv.dtos.BlobStorageNames
@@ -347,5 +346,39 @@ internal class Unit_FnDecompressor {
         var list = eventMap.getOrDefault(hubName, mutableListOf());
         eventMap.put(hubName, list)
         list.addAll(messages);
+    }
+
+    
+    fun defaultMap(id:String = "DUMMY_ID",url:String = "DUMMY_URL", contentType:String = "DUMMY_CONTENT_TYPE"): MutableMap<*,*>{
+        val mapData = mutableMapOf("url" to url, "contentType" to contentType, "extraField" to "DUMMY_EXTRA_FIELD");
+        return mutableMapOf("eventType" to "Microsoft.Storage.BlobCreated", "id" to id, "data" to mapData, "extraField" to "DUMMY_EXTRA_FIELD");
+    }
+
+    fun buildTestMessage(mapObj:Map<*,*>): String { 
+        val message = recursiveBuildMessageFromMap(mapObj);
+        return "[$message]"
+    }
+
+    private fun recursiveBuildMessageFromMap(map:Map<*,*>):String{
+        return map.map { (k,v)-> 
+            var value = if(v is Map<*,*>){recursiveBuildMessageFromMap(v)}else{"\"$v\""};
+            var combined="\"$k\":$value";
+            combined
+        }
+        .joinToString(separator=",",prefix="{",postfix="}") 
+    }
+
+    fun mockContext(): ExecutionContext{
+        val mockContext : ExecutionContext = Mockito.mock(ExecutionContext::class.java);
+        val logger : Logger = Mockito.mock(Logger::class.java);
+
+        Mockito.`when`(mockContext.logger).thenReturn(logger)
+        Mockito.`when`(logger.info(Mockito.anyString())).thenAnswer(::loggerInvocation)
+        return mockContext
+    }
+
+    private fun loggerInvocation(i: InvocationOnMock){
+        val toLog:String = i.getArgument(0);
+        println("[log] $toLog");
     }
 }
